Automate code coverage (#165)
* Add in testing of synapse service module

* Test code coverage

* Fix

* Add init

* Update

* Remove

* Add inits

* Add inits

* Remove

* Try uploading artifact

* Add in flake8

* install all dependencies

* Rename to build

* Remove duplicated test func
