+++ tests/test_challenge.py
@@ -11,16 +11,6 @@ from challengeutils.synapseservices.challenge import Challenge
 from challengeutils.challenge import ChallengeApi
 
 
-def test_challenge_obj():
-    """Tests that a challenge object can be instantiated"""
-    challenge_dict = {'id': 'challenge_1',
-                      'projectId': 'project_1',
-                      'participantTeamId': 'team_1',
-                      'etag': 'etag_1'}
-    challenge_obj = Challenge(**challenge_dict)
-    assert challenge_obj.to_dict() == challenge_dict
-
-
 class TestChallengeApi:
     """Tests ChallengeApi class"""
     def setup(self):
